python/requirements.txt
flask
requests==2.32.3
python-dotenv
cryptography
certifi==2024.8.30
charset-normalizer==3.4.0
idna==3.10
urllib3==2.2.3
actualpy
pandas